+++ NEWS.md
@@ -2,11 +2,12 @@
 News
 ====
 
-Version 0.15a1 - dev
+Version 0.15a2 - dev
 --------------------
 
 - Release notes: https://ezdxf.mozman.at/release-v0-15.html
 - NEW: linetype support for matplotlib drawing backend
+- NEW: HATCH island support for matplotlib drawing backend
 - NEW: recover misplaced tags of the `AcDbEntity` subclass (color, layer, 
   linetype, ...), supported by all loading modes
 - NEW: `ezdxf.addons.geo` module, support for the